backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
numpy==1.26.3
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
aiofiles==23.2.1
